Flutter Driver - Create widget finders from serialized finders extensions
