+++ packages/flutter_driver/lib/src/extension/extension.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import 'package:flutter_driver/src/common/create_finder_factory.dart';
 import 'package:meta/meta.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
@@ -169,7 +170,7 @@ abstract class FinderExtension {
 /// This is not normally used directly. It is instantiated automatically when
 /// calling [enableFlutterDriverExtension].
 @visibleForTesting
-class FlutterDriverExtension with DeserializeFinderFactory {
+class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory {
   /// Creates an object to manage a Flutter Driver connection.
   FlutterDriverExtension(
     this._requestDataHandler,
@@ -354,112 +355,39 @@ class FlutterDriverExtension with DeserializeFinderFactory {
     return finder;
   }
 
-  Finder _createByTextFinder(ByText arguments) {
-    return find.text(arguments.text);
-  }
-
-  Finder _createByTooltipMessageFinder(ByTooltipMessage arguments) {
-    return find.byElementPredicate((Element element) {
-      final Widget widget = element.widget;
-      if (widget is Tooltip)
-        return widget.message == arguments.text;
-      return false;
-    }, description: 'widget with text tooltip "${arguments.text}"');
-  }
-
-  Finder _createBySemanticsLabelFinder(BySemanticsLabel arguments) {
-    return find.byElementPredicate((Element element) {
-      if (element is! RenderObjectElement) {
-        return false;
-      }
-      final String semanticsLabel = element.renderObject?.debugSemantics?.label;
-      if (semanticsLabel == null) {
-        return false;
-      }
-      final Pattern label = arguments.label;
-      return label is RegExp
-          ? label.hasMatch(semanticsLabel)
-          : label == semanticsLabel;
-    }, description: 'widget with semantic label "${arguments.label}"');
-  }
-
-  Finder _createByValueKeyFinder(ByValueKey arguments) {
-    switch (arguments.keyValueType) {
-      case 'int':
-        return find.byKey(ValueKey<int>(arguments.keyValue as int));
-      case 'String':
-        return find.byKey(ValueKey<String>(arguments.keyValue as String));
-      default:
-        throw 'Unsupported ByValueKey type: ${arguments.keyValueType}';
+  @override
+  SerializableFinder deserializeFinder(Map<String, String> json) {
+    final SerializableFinder standard = super.deserializeFinder(json);
+    if (standard != null) {
+      return standard;
     }
-  }
 
-  Finder _createByTypeFinder(ByType arguments) {
-    return find.byElementPredicate((Element element) {
-      return element.widget.runtimeType.toString() == arguments.type;
-    }, description: 'widget with runtimeType "${arguments.type}"');
-  }
-
-  Finder _createPageBackFinder() {
-    return find.byElementPredicate((Element element) {
-      final Widget widget = element.widget;
-      if (widget is Tooltip)
-        return widget.message == 'Back';
-      if (widget is CupertinoNavigationBarBackButton)
-        return true;
-      return false;
-    }, description: 'Material or Cupertino back button');
-  }
-
-  Finder _createAncestorFinder(Ancestor arguments) {
-    final Finder finder = find.ancestor(
-      of: _createFinder(arguments.of),
-      matching: _createFinder(arguments.matching),
-      matchRoot: arguments.matchRoot,
-    );
-    return arguments.firstMatchOnly ? finder.first : finder;
-  }
-
-  Finder _createDescendantFinder(Descendant arguments) {
-    final Finder finder = find.descendant(
-      of: _createFinder(arguments.of),
-      matching: _createFinder(arguments.matching),
-      matchRoot: arguments.matchRoot,
-    );
-    return arguments.firstMatchOnly ? finder.first : finder;
-  }
-
-  Finder _createFinder(SerializableFinder finder) {
-    switch (finder.finderType) {
-      case 'ByText':
-        return _createByTextFinder(finder as ByText);
-      case 'ByTooltipMessage':
-        return _createByTooltipMessageFinder(finder as ByTooltipMessage);
-      case 'BySemanticsLabel':
-        return _createBySemanticsLabelFinder(finder as BySemanticsLabel);
-      case 'ByValueKey':
-        return _createByValueKeyFinder(finder as ByValueKey);
-      case 'ByType':
-        return _createByTypeFinder(finder as ByType);
-      case 'PageBack':
-        return _createPageBackFinder();
-      case 'Ancestor':
-        return _createAncestorFinder(finder as Ancestor);
-      case 'Descendant':
-        return _createDescendantFinder(finder as Descendant);
-      default:
-        if (_finderExtensions.containsKey(finder.finderType)) {
-          return _finderExtensions[finder.finderType].createFinder(finder);
-        } else {
-          throw 'Unsupported finder type: ${finder.finderType}';
-        }
+    final String finderType = json['finderType'];
+    if (_finderExtensions.containsKey(finderType)) {
+      return _finderExtensions[finderType].deserialize(json, this);
     }
+
+    throw DriverError('Unsupported search specification type $finderType');
+  }
+
+  @override
+  Finder createFinder(SerializableFinder finder) {
+    final Finder standard = super.createFinder(finder);
+    if(standard != null) {
+      return standard;
+    }
+
+    if (_finderExtensions.containsKey(finder.finderType)) {
+      return _finderExtensions[finder.finderType].createFinder(finder);
+    }
+
+    throw DriverError('Unsupported search specification type ${finder.finderType}');
   }
 
   Future<TapResult> _tap(Command command) async {
     final Tap tapCommand = command as Tap;
     final Finder computedFinder = await _waitForElement(
-      _createFinder(tapCommand.finder).hitTestable()
+      createFinder(tapCommand.finder).hitTestable()
     );
     await _prober.tap(computedFinder);
     return const TapResult();
@@ -467,13 +395,13 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<WaitForResult> _waitFor(Command command) async {
     final WaitFor waitForCommand = command as WaitFor;
-    await _waitForElement(_createFinder(waitForCommand.finder));
+    await _waitForElement(createFinder(waitForCommand.finder));
     return const WaitForResult();
   }
 
   Future<WaitForAbsentResult> _waitForAbsent(Command command) async {
     final WaitForAbsent waitForAbsentCommand = command as WaitForAbsent;
-    await _waitForAbsentElement(_createFinder(waitForAbsentCommand.finder));
+    await _waitForAbsentElement(createFinder(waitForAbsentCommand.finder));
     return const WaitForAbsentResult();
   }
 
@@ -528,7 +456,7 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<GetSemanticsIdResult> _getSemanticsId(Command command) async {
     final GetSemanticsId semanticsCommand = command as GetSemanticsId;
-    final Finder target = await _waitForElement(_createFinder(semanticsCommand.finder));
+    final Finder target = await _waitForElement(createFinder(semanticsCommand.finder));
     final Iterable<Element> elements = target.evaluate();
     if (elements.length > 1) {
       throw StateError('Found more than one element with the same ID: $elements');
@@ -547,7 +475,7 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<GetOffsetResult> _getOffset(Command command) async {
     final GetOffset getOffsetCommand = command as GetOffset;
-    final Finder finder = await _waitForElement(_createFinder(getOffsetCommand.finder));
+    final Finder finder = await _waitForElement(createFinder(getOffsetCommand.finder));
     final Element element = finder.evaluate().single;
     final RenderBox box = element.renderObject as RenderBox;
     Offset localPoint;
@@ -574,7 +502,7 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<DiagnosticsTreeResult> _getDiagnosticsTree(Command command) async {
     final GetDiagnosticsTree diagnosticsCommand = command as GetDiagnosticsTree;
-    final Finder finder = await _waitForElement(_createFinder(diagnosticsCommand.finder));
+    final Finder finder = await _waitForElement(createFinder(diagnosticsCommand.finder));
     final Element element = finder.evaluate().single;
     DiagnosticsNode diagnosticsNode;
     switch (diagnosticsCommand.diagnosticsType) {
@@ -593,7 +521,7 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<ScrollResult> _scroll(Command command) async {
     final Scroll scrollCommand = command as Scroll;
-    final Finder target = await _waitForElement(_createFinder(scrollCommand.finder));
+    final Finder target = await _waitForElement(createFinder(scrollCommand.finder));
     final int totalMoves = scrollCommand.duration.inMicroseconds * scrollCommand.frequency ~/ Duration.microsecondsPerSecond;
     final Offset delta = Offset(scrollCommand.dx, scrollCommand.dy) / totalMoves.toDouble();
     final Duration pause = scrollCommand.duration ~/ totalMoves;
@@ -614,14 +542,14 @@ class FlutterDriverExtension with DeserializeFinderFactory {
 
   Future<ScrollResult> _scrollIntoView(Command command) async {
     final ScrollIntoView scrollIntoViewCommand = command as ScrollIntoView;
-    final Finder target = await _waitForElement(_createFinder(scrollIntoViewCommand.finder));
+    final Finder target = await _waitForElement(createFinder(scrollIntoViewCommand.finder));
     await Scrollable.ensureVisible(target.evaluate().single, duration: const Duration(milliseconds: 100), alignment: scrollIntoViewCommand.alignment ?? 0.0);
     return const ScrollResult();
   }
 
   Future<GetTextResult> _getText(Command command) async {
     final GetText getTextCommand = command as GetText;
-    final Finder target = await _waitForElement(_createFinder(getTextCommand.finder));
+    final Finder target = await _waitForElement(createFinder(getTextCommand.finder));
 
     final Widget widget = target.evaluate().single.widget;
     String text;
